// server/prisma/data.d.ts
export declare const employees: {
    firstName: string;
    lastName: string;
    email: string;
    onboardee: boolean;
    team: string;
    role: string;
    level: string;
    status: string;
    picture: string;
    skills: {
        rating: number;
        skill: {
            name: string;
        };
    }[];
    interests: {
        name: string;
        tags: {
            name: string;
        }[];
    }[];
    groups: {
        name: string;
        tags: {
            name: string;
        }[];
    }[];
}[];